Remove unnecessary wrapper class InternalHTTPRequest
